TCA-48 The mappers were create and some file moved from fold
