+++ app/src/main/java/ar/scacchipa/twittercloneapp/domain/ConsumableAuthUseCase.kt
@@ -1,6 +1,6 @@
 package ar.scacchipa.twittercloneapp.domain
 
-import ar.scacchipa.twittercloneapp.repository.Constants
+import ar.scacchipa.twittercloneapp.utils.Constants
 
 open class ConsumableAuthUseCase {
     open operator fun invoke(): String {
