+++ app/src/main/java/ar/scacchipa/twittercloneapp/fragment/FragmentAuthWebDialog.kt
@@ -12,7 +12,7 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
 import ar.scacchipa.twittercloneapp.databinding.FragmentAuthWebDialogLayoutBinding
-import ar.scacchipa.twittercloneapp.repository.TokenResource
+import ar.scacchipa.twittercloneapp.data.TokenResource
 import ar.scacchipa.twittercloneapp.viewmodel.AuthWebDialogViewModel
 import java.net.URI
