+++ app/src/main/java/ar/scacchipa/twittercloneapp/repository/IAuthorizationRepository.kt
@@ -1,5 +1,7 @@
 package ar.scacchipa.twittercloneapp.repository
 
+import ar.scacchipa.twittercloneapp.data.TokenResource
+
 interface IAuthorizationRepository {
     suspend fun requestAccessToken(
         transitoryToken: String,
